consjuncitve support with dual ok

+++ scripts/launch.py
@@ -8,7 +8,7 @@ from shutil import copyfile
 import fileinput
 import compil as cp
 
-def run_miner(dataset_folder, cpp_log_file, minimal_option):
+def run_miner(dataset_folder, cpp_log_file, minimal_option, consjonctive_option):
     onlyfiles = [f for f in os.listdir(dataset_folder) if os.path.isfile(os.path.join(dataset_folder, f))]
     if(len(onlyfiles) == 0):
         print("no dataset in ", dataset_folder)
@@ -28,7 +28,7 @@ def run_miner(dataset_folder, cpp_log_file, minimal_option):
 
                 os.chdir("../MT-Miner")
                 start = perf_counter()            
-                cp.compil_and_run_miner(fullpath, cpp_log_file, minimal_option)
+                cp.compil_and_run_miner(fullpath, cpp_log_file, minimal_option, consjonctive_option)
                 #time.sleep(0.25)
                 end = perf_counter()
                 print("Elapsed time for compilation and Miner execution: ", end - start, " sec")
@@ -36,12 +36,18 @@ def run_miner(dataset_folder, cpp_log_file, minimal_option):
 
         os.chdir(current_path)
 
-        df_cpp = pd.read_csv(cpp_log_file, delimiter=';', header=None)
-        df_cpp.columns = ['filename', 'sparcity', 'cloneCount', 'itemCount', 'objectCount', 'minerMinimalTransverseCount', 'minimalSizeOfTransverse', 'minerTime']
-        df_cpp.index = df_cpp["filename"]
-        df_cpp.index.name = "filename"
-        #df = df.drop(["filename"], axis=1)
-        df_cpp.to_csv(cpp_log_file, sep=';')
+        if os.path.isfile(cpp_log_file):
+            df_cpp = pd.read_csv(cpp_log_file, delimiter=';', header=None)
+            if minimal_option == "false":
+                df_cpp.columns = ['filename', 'sparcity', 'cloneCount', 'itemCount', 'objectCount', 'minerMinimalTransverseCount', 'minimalSizeOfTransverse', 'minerTime']
+            else:
+                df_cpp.columns = ['filename', 'sparcity', 'cloneCount', 'itemCount', 'objectCount', 'minerMinimalTransverseCount', 'MinimalSizeOfTransverse', 'minerTime']
+            df_cpp.index = df_cpp["filename"]
+            df_cpp.index.name = "filename"
+            #df = df.drop(["filename"], axis=1)
+            df_cpp.to_csv(cpp_log_file, sep=';')
+        else:
+            print("no data found")
 
 
 def run_shd(dataset_folder, shd_log_file):
@@ -84,6 +90,7 @@ def run_shd(dataset_folder, shd_log_file):
         df_shd.index.name = "filename"
         df_shd.to_csv(shd_log_file, sep=';')
 
+
 def run_pmmcs(dataset_folder, log_file):
     onlyfiles = [f for f in os.listdir(dataset_folder) if os.path.isfile(os.path.join(dataset_folder, f))]
     if(len(onlyfiles) == 0):
@@ -105,8 +112,7 @@ def run_pmmcs(dataset_folder, log_file):
                 print("pMMCS execution...")                    
                 os.chdir("../Minimal-Hitting-Set-Algorithms")
                 start = perf_counter()
-                res = subprocess.run(["./agdmhs", fullpath, "out.dat", "-a", "pmmcs"], stdout=subprocess.PIPE)
-                
+                res = subprocess.run(["./agdmhs", fullpath, "out.dat", "-a", "pmmcs"], stdout=subprocess.PIPE)                
                 cpt = 0
                 with fileinput.FileInput("out.dat") as file:
                     for line in file:
@@ -129,11 +135,57 @@ def run_pmmcs(dataset_folder, log_file):
         df_pmmcs.to_csv(log_file, sep=';')
 
 
+def run_prs(dataset_folder, log_file):
+    onlyfiles = [f for f in os.listdir(dataset_folder) if os.path.isfile(os.path.join(dataset_folder, f))]
+    if(len(onlyfiles) == 0):
+        print("no dataset in ", dataset_folder)
+    else:        
+        current_path = os.getcwd()
+
+        log_file = os.path.join(current_path, log_file)
+        if os.path.isfile(log_file):
+            os.remove(log_file)
+
+        data_pmmcs = []  
+        for file in onlyfiles:
+            if file.lower().endswith(('.txt', '.dat')):
+                fullpath = os.path.join(dataset_folder, file)
+
+                print("processing data file:", fullpath)
+
+                print("pMMCS execution...")                    
+                os.chdir("../Minimal-Hitting-Set-Algorithms")
+                start = perf_counter()
+                res = subprocess.run(["./agdmhs", fullpath, "out.dat", "-a", "prs"], stdout=subprocess.PIPE)                
+                cpt = 0
+                with fileinput.FileInput("out.dat") as file:
+                    for line in file:
+                        cpt = cpt+1
+
+                #time.sleep(0.5)
+                end = perf_counter()
+                total = end - start
+                print("Elapsed time for pMMCS execution:", total, " sec")
+                print("\n")
+                    
+                # append shd value
+                data_pmmcs.append([fullpath, cpt, total])        
+        
+        os.chdir(current_path)
+
+        df_pmmcs = pd.DataFrame(data_pmmcs, columns=['filename', 'prsMinimalTransverseCount', 'prsSTime'])
+        df_pmmcs.index = df_pmmcs["filename"]
+        df_pmmcs.index.name = "filename"
+        df_pmmcs.to_csv(log_file, sep=';')
+
+
 print("usage :  python3 ./launch.py \"../data/\" <all/all_miner/min_miner/shd>")
 print("<all> : compute all minimal transverses with miner, then compute only minimal transverses with minimal size with miner, then compute all minimal transverses with shd")
 print("<all_miner> : compute all minimal transverses with miner only")
 print("<min_miner> : compute only minimal transverses with minimal size with miner only")
 print("<shd> : compute all minimal transverses with shd only")
+print("<pmmcs> : compute all minimal transverses with pmmcs only")
+print("<prs> : compute all minimal transverses with prs only")
 
 if(len(sys.argv) != 3):
     print("usage :  python3 ./launch.py \"../data/\" <all/all_miner/min_miner/shd>")
@@ -143,25 +195,48 @@ else:
 
     if option == "all" or option == "all_miner":
         print("compute all minimal transverses with miner")
-        run_miner(dataset_folder, "cpp_log_all.csv", "false")
-        output = dataset_folder + "cpp_log_all.csv"
-        copyfile("cpp_log_all.csv", output)
+        log_file = "miner_log_all.csv"
+        run_miner(dataset_folder, log_file, "false", "false")
+        output = dataset_folder + log_file
+        if os.path.isfile(output):
+            copyfile(log_file, output)
 
     if option == "all" or option == "min_miner":
         print("compute only minimal transverses with minimal size with miner")
-        run_miner(dataset_folder, "cpp_log_cut.csv", "true")
-        output = dataset_folder + "cpp_log_all.csv"
-        copyfile("cpp_log_cut.csv", output)
+        log_file = "miner_log_cut.csv"
+        run_miner(dataset_folder, log_file, "true", "false")
+        output = dataset_folder + log_file
+        if os.path.isfile(output):
+            copyfile(log_file, output)
+
+    if option == "all" or option == "all_miner":
+        print("compute all minimal transverses with miner")
+        log_file = "miner_log_all_consjonctive.csv"
+        run_miner(dataset_folder, log_file, "false", "true")
+        output = dataset_folder + log_file
+        if os.path.isfile(output):
+            copyfile(log_file, output)
 
     if option == "all" or option == "shd":    
         print("compute all minimal transverses with shd")
-        run_shd(dataset_folder, "shd_log_file.csv")
-        output = dataset_folder + "shd_log_file.csv"   
-        copyfile("shd_log_file.csv", output) 
+        log_file = "shd_log_file.csv"
+        run_shd(dataset_folder, log_file)
+        output = dataset_folder + log_file
+        if os.path.isfile(output):
+            copyfile(log_file, output) 
 
     if option == "all" or option == "pmmcs":
         print("compute all minimal transverses with pMMCS")
-        run_pmmcs(dataset_folder, "pmmcs_log_file.csv")
-        output = dataset_folder + "pmmcs_log_file.csv"   
-        copyfile("pmmcs_log_file.csv", output) 
+        log_file = "pmmcs_log_file.csv"
+        run_pmmcs(dataset_folder, log_file)
+        output = dataset_folder + log_file
+        if os.path.isfile(output):
+            copyfile(log_file, output) 
 
+    if option == "all" or option == "prs":
+        print("compute all minimal transverses with prs")
+        log_file = "prs_log_file.csv"
+        run_pmmcs(dataset_folder, log_file)
+        output = dataset_folder + log_file
+        if os.path.isfile(output):
+            copyfile(log_file, output)
